Release 1.21.00 - rebase advanced branch to current trunk
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1851 98333e67-4a24-44d7-a75c-e53540dd3050

+++ advanced/src/test/value.c
@@ -1421,7 +1421,7 @@ test_struct_get_element(xmlrpc_value * const structP,
 
 static void
 testStructReadout(xmlrpc_value * const structP,
-                  size_t         const expectedSize) {
+                  unsigned int   const expectedSize) {
 
     xmlrpc_env env;
     xmlrpc_value * keyP;
